requirements.txt
Flask
gunicorn
requests
twilio
azure-identity
azure-keyvault-secrets